% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/defaultPar.R
\name{defaultPar}
\alias{defaultPar}
\title{Default parameters}
\usage{
defaultPar(mu, n = 20, mixed = TRUE)
}
\arguments{
\item{mu}{vector of paramters}

\item{n}{number of series (if mixed=TRUE)}

\item{mixed}{default=TRUE; if false: \eqn{\phi_i=\mu}}
}
\value{
\item{phi}{example values of \eqn{\phi}}
\item{mu}{example values of \eqn{\mu}}
\item{Omega}{example values of \eqn{\Omega}}
\item{gamma2}{example values of \eqn{\gamma^2}}
\item{t}{example time points}
\item{startY}{starting value for the SDE}
}
\description{
Yields example parameters conditional on \eqn{\mu} for the different models.
}

