% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drawData.R
\name{drawData}
\alias{drawData}
\title{Function for simulating the data}
\usage{
drawData(model, fun, para, sVar, mw = 10)
}
\arguments{
\item{model}{one out of {SDE, ODE}}

\item{fun}{regression or drift function}

\item{para}{list of parameters, for example defaultPar(getPar(model, para=truePar))}

\item{sVar}{variance function}

\item{mw}{mesh width (in the case of SDE)}
}
\value{
data series in para$t
}
\description{
Simulation of variables depend on the model.
}

