% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estReg.R
\name{estReg}
\alias{estReg}
\title{Bayesian estimation in mixed nonlinear regression models}
\usage{
estReg(t, y, prior, start, fODE, sVar, ipred = 1, cut, len = 1000,
  Omega = "diag", mod = c("Gompertz", "logistic", "Weibull", "Richards",
  "Paris", "Paris2"), propPar = 0.02)
}
\arguments{
\item{t}{vector of observation times}

\item{y}{matrix of the M trajectories}

\item{prior}{list of prior parameters - list(m, v, priorOmega, alpha, beta)}

\item{start}{list of starting values}

\item{fODE}{regression function}

\item{sVar}{variance function}

\item{ipred}{which of the M trajectories is the one to be predicted}

\item{cut}{the index how many of the ipred-th series are used for estimation}

\item{len}{number of iterations of the MCMC algorithm}

\item{Omega}{structure of the variance matrix Omega of the random effects, diagonal matrix, otherwise inverse wishart distributed}

\item{mod}{model out of {Gompertz, Richards, logistic, Weibull}, only used instead of fODE}

\item{propPar}{proposal standard deviation of phi is |start$mu|*propPar}
}
\value{
\item{phi}{samples from posterior of \eqn{\phi}}
\item{mu}{samples from posterior of \eqn{\mu}}
\item{Omega}{samples from posterior of \eqn{\Omega}}
\item{gamma2}{samples from posterior of \eqn{\gamma^2}}
}
\description{
Bayesian estimation of the random effects \eqn{\phi_j} in the mixed nonlinear regression model
 \eqn{y_{ij}= f(\phi_j, t_{ij}) + \epsilon_{ij}, \epsilon_{ij}~N(0,\gamma^2*s^2(t_{ij}), \phi_j~N(\mu, \Omega)} and the parameters
 \eqn{\mu, \Omega, \gamma^2}.
}
\examples{
mod <- "Gompertz"
fODE <- getFun("ODE", mod)
mu <- getPar("ODE", mod, "truePar")
n <- 5
parameters <- defaultPar(mu, n)
y <- drawData("ODE", fODE, parameters)
t <- parameters$t
prior <- getPrior(mu, parameters$gamma2)
start <- getStart(mu, n)
chains <- estReg(t, y, prior=prior, start=start, fODE=fODE)
plot(phi_ij(chains$phi, 1, 1), type="l")
plot(chains$gamma, type="l"); abline(h=parameters$gamma2, col=2)
}

