% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estReg_single.R
\name{estReg_single}
\alias{estReg_single}
\title{Bayesian stimation in nonlinear regression models}
\usage{
estReg_single(t, y, prior, start, fODE, sVar, len = 1000, mod = "Gompertz")
}
\arguments{
\item{t}{vector of observation times}

\item{y}{vector of the M trajectories}

\item{prior}{list of prior parameters - list(mu, Omega, alpha, beta)}

\item{start}{list of starting values}

\item{fODE}{regression function}

\item{sVar}{variance function}

\item{len}{number of iterations of the MCMC algorithm}

\item{mod}{model out of {Gompertz, Richards, logistic, Weibull}, only used instead of fODE}

\item{propSd}{proposal standard deviation of phi is |mu|*propSd}
}
\value{
\item{phi}{estimator of \eqn{\phi}}
\item{gamma2}{estimator of \eqn{\gamma^2}}
}
\description{
Bayesian estimation of the parameters of the mixed nonlinear regression model
 \eqn{y_j= f(\phi, t_j) + \epsilon_j, \epsilon_j~N(0,\gamma^2*s^2(t_j)}.
}

