% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estSDE.R
\name{estSDE}
\alias{estSDE}
\title{Bayesian estimation in mixed stochastic differential equations}
\usage{
estSDE(t, y, prior, start, bSDE, sVar, ipred = 1, cut, len = 1000,
  Omega = "diag", mod = c("Gompertz", "logistic", "Weibull", "Richards",
  "Paris", "Paris2"), modVar = "", propPar = 0.02)
}
\arguments{
\item{t}{vector of observation times}

\item{y}{matrix or list of the n trajectories}

\item{prior}{list of prior parameters - list(m, v, priorOmega, alpha, beta), priorOmega=list(alpha, beta) if Omega="diag", otherwise prior matrix of Wishart distribution}

\item{start}{list of starting values}

\item{bSDE}{b(phi, t, x) drift function}

\item{sVar}{variance function s^2}

\item{ipred}{which of the n trajectories is the one to be predicted}

\item{cut}{the index how many of the ipred-th series are used for estimation}

\item{len}{number of iterations of the MCMC algorithm - chain length}

\item{Omega}{structure of the variance matrix Omega of the random effects, diagonal matrix, otherwise inverse wishart distributed}

\item{mod}{model out of {Gompertz, Richards, logistic, Weibull, Paris, Paris2}, only used instead of bSDE}

\item{modVar}{default value is sVar(t,x)=1, if "AR": sVar(t,x)=x}

\item{propPar}{proposal standard deviation of phi is |start$mu|*propPar}
}
\value{
\item{phi}{samples from posterior of \eqn{\phi}}
\item{mu}{samples from posterior of \eqn{\mu}}
\item{Omega}{samples from posterior of \eqn{\Omega}}
\item{gamma2}{samples from posterior of \eqn{\gamma^2}}
}
\description{
Bayesian estimation of the random effects \eqn{\phi_j} in the mixed SDE
 \eqn{dY_i(t)= b(\phi_i, t, Y_i(t))dt + \gamma s(t, Y_i(t)) dW_i(t), , \phi_i~N(\mu, \Omega), i=1,...,n} and the parameters
 \eqn{\mu, \Omega, \gamma^2}.
}
\details{
Simulation from the posterior distribution of the random effect from n independent trajectories of the SDE (the Brownian motions \eqn{W1,...,Wn} are independent).
\subsection{?}{
}
}
\examples{
mod <- "Gompertz"
bSDE <- getFun("SDE", mod)
mu <- getPar("SDE", mod, "truePar")
n <- 5
parameters <- defaultPar(mu, n)
y <- drawData("SDE", bSDE, parameters)
t <- parameters$t

prior <- getPrior(mu, parameters$gamma2)
start <- getStart(mu, n)
chains <- estSDE(t, y, prior, start, bSDE)
}
\references{
Hermann et al. (2015)
}
\keyword{estimation}

