% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estSDE_single.R
\name{estSDE_single}
\alias{estSDE_single}
\title{Bayesian stimation in mixed stochastic differential equations}
\usage{
estSDE_single(t, X, prior, start, bSDE, sVar, len = 1000, mod = "Gompertz",
  modVar = "")
}
\arguments{
\item{t}{vector of observation times}

\item{X}{vector of the M trajectories}

\item{prior}{list of prior parameters - list(mu, Omega, alpha, beta)}

\item{start}{list of starting values}

\item{sVar}{variance function}

\item{len}{number of iterations of the MCMC algorithm}

\item{mod}{model out of {Gompertz, Richards, logistic, Weibull}, only used instead of fODE}

\item{modVar}{default value is sVar(t,x)=1, if "AR": sVar(t,x)=x}

\item{fODE}{regression function}

\item{propSd}{proposal standard deviation of phi is |mu|*propSd}
}
\value{
\item{phi}{estimator of \eqn{\phi}}
\item{gamma2}{estimator of \eqn{\gamma^2}}
}
\description{
Bayesian estimation of the parameters in the mixed SDE
 \eqn{dY(t)= b(\phi, t, Y(t))dt + \gamma s(t, Y(t)) dW(t)}.
}

