% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFun.R
\name{getFun}
\alias{getFun}
\title{Yields the regression or drift function for the specific models}
\usage{
getFun(model, mod)
}
\arguments{
\item{model}{one out of {SDE, ODE}}

\item{mod}{one out of {Gompertz, Richards, logistic, Weibull}}
}
\value{
function
}
\description{
Gives out the regression function (ODE) or the drift function (SDE) for the specific model out of {Gompertz, Richards, logistic, Weibull}
}

