% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFunVar.R
\name{getFunVar}
\alias{getFunVar}
\title{Variance function}
\usage{
getFunVar(modelVar = "")
}
\arguments{
\item{modelVar}{one out of {, AR}}
}
\value{
function
}
\description{
Gives out the variance function (SDE)
}

