% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getPar.R
\name{getPar}
\alias{getPar}
\title{Default values}
\usage{
getPar(model = "ODE", mod = "Gompertz", para = c("prior", "truePar"))
}
\arguments{
\item{model}{one out of {SDE, ODE}}

\item{mod}{one out of {Gompertz, Richards, logistic, Weibull}}

\item{para}{prior values for the data set of Virkler et al. (1979) or parameters for the simultions}
}
\value{
vector of parameters
}
\description{
Default values for \eqn{\mu} for the specific model out of {Gompertz, Richards, logistic, Weibull}
}

