% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getPrior.R
\name{getPrior}
\alias{getPrior}
\title{Builds list of prior parameters}
\usage{
getPrior(mu, gamma2, mixed = TRUE)
}
\arguments{
\item{mu}{}

\item{gamma2}{}

\item{mixed}{}
}
\value{
list of prior values
}
\description{
Creation of list of parameters conditional on true values \eqn{\mu} and \eqn{\gamma^2}
}

