% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getStart.R
\name{getStart}
\alias{getStart}
\title{Builds list of starting values}
\usage{
getStart(mu, n = 1, mixed = TRUE)
}
\arguments{
\item{mu}{}

\item{n}{}

\item{mixed}{}
}
\value{
list of starting values
}
\description{
Creation of list of parameters conditional on true value \eqn{\mu}
}

