% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phi_ij.R
\name{phi_ij}
\alias{phi_ij}
\title{Helping function}
\usage{
phi_ij(phi, i, j)
}
\arguments{
\item{phi}{list of samples for mixed effect}

\item{i}{component of vector}

\item{j}{which individual}
}
\value{
samples for the i-th component of \eqn{\phi_j}
}
\description{
transfers the list to one vector
}

