% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postmu.R
\name{postmu}
\alias{postmu}
\title{Posterior for mu}
\usage{
postmu(phi, m, v, Omega)
}
\arguments{
\item{phi}{}

\item{v}{}

\item{Omega}{}

\item{mu}{}
}
\value{
one sample of posterior
}
\description{
Posterior for parameters \eqn{\mu}
}

