% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postOmega.R
\name{postOmega}
\alias{postOmega}
\title{Posterior}
\usage{
postOmega(R, phi, mu)
}
\arguments{
\item{R}{}

\item{phi}{}

\item{mu}{}
}
\value{
one sample of posterior
}
\description{
Posterior for parameters \eqn{\Omega}
}

