% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postOmegaDiag.R
\name{postOmegaDiag}
\alias{postOmegaDiag}
\title{Posterior}
\usage{
postOmegaDiag(alpha, beta, phi, mu)
}
\arguments{
\item{alpha}{}

\item{beta}{}

\item{phi}{}

\item{mu}{}
}
\value{
one sample of posterior
}
\description{
Posterior for parameters \eqn{\Omega}
}

