% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predPhi.R
\name{predPhi}
\alias{predPhi}
\title{Prediction Of The Random Effects In Mixed Stochastic Differential Equations}
\usage{
predPhi(samples, cand)
}
\arguments{
\item{samples}{output of estSDE or estREg}

\item{cand}{candidates for phi (matrix with p columns)}
}
\value{
matrix phi
}
\description{
Prediction of the random effects
}
\examples{
mod <- "Gompertz"
bSDE <- getFun("SDE", mod)
mu <- getPar("SDE", mod, "truePar")
n <- 5
parameters <- defaultPar(mu, n)
y <- drawData("SDE", bSDE, parameters)
t <- parameters$t

prior <- getPrior(mu, parameters$gamma2)
start <- getStart(mu, n)
chains <- estSDE(t, y, prior, start, bSDE)
phi.pred <- predPhi(chains)
}
\references{
follows...
}
\keyword{prediction}

