% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predReg.R
\name{predReg}
\alias{predReg}
\title{Bayesian prediction in mixed nonlinear regression models}
\usage{
predReg(t, samples, fODE, sVar, cand, len = 1000, mod = "Gompertz")
}
\arguments{
\item{t}{vector of times which are predicted}

\item{samples}{list of samples from the posterior}

\item{fODE}{regression function}

\item{sVar}{variance function}

\item{cand}{vector of candidates for trajection sampling}

\item{len}{number of samples from the predictive distribution}

\item{mod}{model out of {Gompertz, Richards, logistic, Weibull}, only used instead of fODE}
}
\value{
matrix of predictions in t
}
\description{
Bayesian prediction in the mixed nonlinear regression model
 \eqn{y_{ij}= f(\phi_j, t_{ij}) + \epsilon_{ij}, \epsilon_{ij}~N(0,\gamma^2*s^2(t_{ij}), \phi_j~N(\mu, \Omega)}.
}

