% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predSDE.R
\name{predSDE}
\alias{predSDE}
\title{Bayesian prediction in mixed SDE models}
\usage{
predSDE(tau, samples, last, bSDE, sVar, cand, mod = "Gompertz", modVar = "")
}
\arguments{
\item{tau}{vector of times which are predicted}

\item{samples}{list of samples from the posterior}

\item{last}{last observation - staring point for the prediction of Markov chain}

\item{bSDE}{drift function}

\item{sVar}{variance function}

\item{cand}{vector of candidates for trajection sampling}

\item{mod}{model out of {Gompertz, Richards, logistic, Weibull}, only used instead of bSDE}

\item{modVar}{model for the variance structure}
}
\value{
matrix of predictions in t
}
\description{
Bayesian prediction in the mixed SDE
 \eqn{dY_j(t)= b(\phi_j, t, Y_j(t))dt + \gamma s(t, Y_j(t)) dW_j(t), , \phi_j~N(\mu, \Omega)}.
}
\examples{
mod <- "Gompertz"
bSDE <- getFun("SDE", mod)
mu <- getPar("SDE", mod, "truePar")
n <- 5
parameters <- defaultPar(mu, n)
y <- drawData("SDE", bSDE, parameters)
t <- parameters$t

prior <- getPrior(mu, parameters$gamma2)
start <- getStart(mu, n)
chains <- estSDE(t, y, prior, cut = 50, ipred = 1, start, bSDE, len = 5000)
ind <- seq(1001, 5000, by = 4)
samples <- list(phi = sapply(1:length(mu), function(i) phi = phi_ij(chains$phi, 1, i))[ind, ], gamma2 = chains$gamma2[ind])
prediction <- predSDE(t[50:101], samples, y[1,50], bSDE, cand = seq(-2, 2, length = 1000))
plot(t[51:101], y[1,51:101], ylim = range(y[1,51:101]) + c(0, 1))
lines(t[51:101], apply(prediction, 1, quantile, 0.025), col = 3)
lines(t[51:101], apply(prediction, 1, quantile, 0.975), col = 3)
lines(t[51:101], apply(prediction, 1, mean), col = 2)
}

