% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scoreRule.R
\name{scoreRule}
\alias{scoreRule}
\title{Scoring rule of Gneiting and Raftery (...)}
\usage{
scoreRule(l, u, x)
}
\arguments{
\item{l}{lower bound}

\item{u}{upper bound}

\item{x}{true value}
}
\value{
interval score
}
\description{
Interval score.
}

