% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testSDE.R
\name{testSDE}
\alias{testSDE}
\title{Example function for the SDE model}
\usage{
testSDE(mod = "Gompertz", mixed = TRUE)
}
\arguments{
\item{mod}{model out of {Gompertz, Richards, logistic, Weibull}}

\item{mixed}{default=TRUE; if false: \eqn{\phi_i=\mu}}
}
\value{
nice pictures
}
\description{
Example function - how to use the functions in the package
}

