% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predSDE.R
\name{predSDE}
\alias{predSDE}
\title{Bayesian prediction in mixed SDE models}
\usage{
predSDE(tau, samples, last, bSDE, sVar, cand, mod = "Gompertz", modVar = "")
}
\arguments{
\item{tau}{vector of times which are predicted}

\item{samples}{list of samples from the posterior}

\item{last}{last observation - staring point for the prediction of Markov chain}

\item{bSDE}{drift function}

\item{sVar}{variance function}

\item{cand}{vector of candidates for trajection sampling}

\item{mod}{model out of {Gompertz, Richards, logistic, Weibull}, only used instead of bSDE}

\item{modVar}{model for the variance structure}
}
\value{
matrix of predictions in t
}
\description{
Bayesian prediction in the mixed SDE
 \eqn{dY_j(t)= b(\phi_j, t, Y_j(t))dt + \gamma s(t, Y_j(t)) dW_j(t), , \phi_j~N(\mu, \Omega)}.
}

